11 Jul. (Sujee) : slight changed

+++ lib/first_tab.dart
@@ -568,8 +568,8 @@ class Letter {
       receivedId: json['receiverId'],
       senderId: json['senderId'],
       text: json['text'],
-      posX: json['posX'],
-      posY: json['posY'],
+      posX: json['posX'].toDouble(),
+      posY: json['posY'].toDouble(),
       imgType: json['imgType']
     );
   }
